Move fake actions under proper pkg
Fake actions are located under kubecli pkg.
This is wrong since it is not related to kubecli,
but to testing in general.
