+++ staging/src/kubevirt.io/client-go/testing/BUILD.bazel
@@ -3,7 +3,7 @@ load("@io_bazel_rules_go//go:def.bzl", "go_library")
 go_library(
     name = "go_default_library",
     srcs = ["actions.go"],
-    importpath = "kubevirt.io/client-go/kubecli/fake",
+    importpath = "kubevirt.io/client-go/testing",
     visibility = ["//visibility:public"],
     deps = [
         "//vendor/k8s.io/apimachinery/pkg/runtime/schema:go_default_library",
