+++ staging/src/kubevirt.io/client-go/testing/actions.go
@@ -17,7 +17,7 @@
  *
  */
 
-package fake
+package testing
 
 import (
 	"k8s.io/apimachinery/pkg/runtime/schema"
